[update] reorganize a bit
